Reorganized sortition/internal local chain implementation
Moved `local.go` to `sortition/internal/local` to keep it in a separate
package from `sortition`. `connectLocal` renamed to `Connect`,
`registerOperator` and `setEligibleStake` are now exported.
